exercises/requirements.txt
#Python 3.10.19
absl-py==2.4.0
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.7
numpy==2.2.6
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pillow==12.1.1
protobuf==3.20.3
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.5
Werkzeug==3.1.6
mysql-connector-python
